user and service

+++ database/migrations/2021_05_15_060538_create_product_metas_table.php
@@ -15,7 +15,7 @@ class CreateProductMetasTable extends Migration
     {
         Schema::create('product_metas', function (Blueprint $table) {
             $table->id();
-            $table->integer('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->string('meta_key');
             $table->text('meta_value');
             $table->timestamps();
